Restricoes de acesso implementadas

+++ src/main/java/br/fatec/financas/service/ContaService.java
@@ -10,9 +10,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import br.fatec.financas.dto.ContaDTO;
+import br.fatec.financas.exception.AuthorizationException;
 import br.fatec.financas.mapper.ContaMapper;
+import br.fatec.financas.model.Cliente;
 import br.fatec.financas.model.Conta;
+import br.fatec.financas.repository.ClienteRepository;
 import br.fatec.financas.repository.ContaRepository;
+import br.fatec.financas.security.JWTUtil;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
@@ -21,9 +25,15 @@ public class ContaService implements ServiceInterface<ContaDTO> {
 
 	@Autowired
 	private ContaRepository repository;
+	
+	@Autowired
+	private ClienteRepository clienteRepo;
 
 	@Autowired
 	private ContaMapper mapper;
+	
+	@Autowired
+	private JWTUtil jwtUtil;
 	
 	@Override
 	public ContaDTO create(ContaDTO conta) {
@@ -49,7 +59,11 @@ public class ContaService implements ServiceInterface<ContaDTO> {
 	}
 	
 	@Override
-	public ContaDTO findById(Long id) {
+	public ContaDTO findById(Long id) throws AuthorizationException {
+		Cliente cliente = clienteRepo.findByConta(id);
+		if (cliente != null && !jwtUtil.authorized(cliente.getId())) {
+			throw new AuthorizationException("Acesso negado!"); 
+		}
 		Optional<Conta> obj = repository.findById(id);
 		if (obj.isPresent()) {
 			return mapper.toDTO(obj.get());
@@ -59,7 +73,11 @@ public class ContaService implements ServiceInterface<ContaDTO> {
 
 	
 	@Override
-	public boolean update(ContaDTO conta) {
+	public boolean update(ContaDTO conta) throws AuthorizationException {
+		Cliente cliente = clienteRepo.findByConta(conta.getId());
+		if (cliente != null && !jwtUtil.authorized(cliente.getId())) {
+			throw new AuthorizationException("Acesso negado!"); 
+		}
 		if (repository.existsById(conta.getId())) {
 			repository.save(mapper.toEntity(conta));
 			return true;
@@ -70,15 +88,23 @@ public class ContaService implements ServiceInterface<ContaDTO> {
 	@Override
 	public boolean delete(Long id) {
 		if (repository.existsById(id)) {
-			repository.deleteById(id);
+			Cliente cliente = clienteRepo.findByConta(id);
+			if (cliente != null) {
+				throw new IllegalArgumentException("Há um cliente vinculado a essa conta!");
+			}
+			repository.deleteById(id);			
 			return true;
 		}
 		return false;
 	}
 
-	public Float depositar(Long id, Float valor) {
+	public Float depositar(Long id, Float valor) throws AuthorizationException {
 		ContaDTO _conta = findById(id);
 		if (_conta != null) {
+			Cliente cliente = clienteRepo.findByConta(_conta.getId());
+			if (cliente != null && !jwtUtil.authorized(cliente.getId())) {
+				throw new AuthorizationException("Acesso negado!"); 
+			}
 			_conta.setSaldo(_conta.getSaldo() + valor);
 			update(_conta);
 			return _conta.getSaldo();
@@ -86,9 +112,13 @@ public class ContaService implements ServiceInterface<ContaDTO> {
 		return null;
 	}
 
-	public Float sacar(Long id, Float valor) throws IllegalArgumentException {
+	public Float sacar(Long id, Float valor) throws IllegalArgumentException, AuthorizationException {
 		ContaDTO _conta = findById(id);
 		if (_conta != null) {
+			Cliente cliente = clienteRepo.findByConta(_conta.getId());
+			if (cliente != null && !jwtUtil.authorized(cliente.getId())) {
+				throw new AuthorizationException("Acesso negado!"); 
+			}
 			Float _saldo = _conta.getSaldo();
 			if (_saldo >= valor) {
 				_conta.setSaldo(_saldo - valor);
